Widen integer literals to 64-bit (were silently truncated to 32-bit)
`ExprLiteral.value` was a 32-bit `int`, and expr_literal/the parser cast
`parse_numeric_literal`'s `long long` down to `int`. So EVERY literal above
2^31-1 was truncated: `var x i64 = 5000000000` became 705032704, and u64
constants above i64 max were garbage. Change the field and constructor to
int64_t, drop the `(int)` casts at the four parser call sites, and emit the
value with %lld. Large literals now round-trip (i64 5000000000 prints
correctly). Updated vra_l2_int_concretize_pass (it relied on the truncation
+ the cut i32->i64 monomorphization). char_i64_literal_pass. 320/320 green.

+++ src/ast.h
@@ -435,7 +435,8 @@ typedef struct {
 } ExprIdentifier;
 
 typedef struct {
-    int         value;  // Literal value (TODO: extend to support other types)
+    int64_t     value;  // Literal integer value (64-bit; was 32-bit int — truncated
+                        // any literal above 2^31-1, e.g. i64/u64 constants).
 } ExprLiteral;
 
 typedef struct {
@@ -1013,7 +1014,7 @@ Expr *expr_identifier(Arena *arena, Id *id) {
     return e;
 }
 
-Expr *expr_literal(Arena *arena, int value) {
+Expr *expr_literal(Arena *arena, int64_t value) {
     Expr *e = arena_push_aligned(arena, Expr);
     e->kind = EXPR_LITERAL;
     e->as.literal_expr.value = value;
