+++ src/ast_print.h
@@ -78,7 +78,7 @@ void expr_print_ast(Expr *expr, int depth) {
     indent(depth);
     switch (expr->kind) {
         case EXPR_LITERAL:
-            printf("Literal: %d\n", expr->as.literal_expr.value);
+            printf("Literal: %lld\n", (long long)expr->as.literal_expr.value);
             break;
         case EXPR_FLOAT_LITERAL:
             printf("Float Literal: %.17g\n", expr->as.float_expr.value);
